% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backwards_selection.R
\name{backward_selection}
\alias{backward_selection}
\title{Backwards Selection}
\usage{
backward_selection(data, y_var, x_vars, significance_level = 0.05)
}
\arguments{
\item{data}{data frame}

\item{y_var}{string of target variable column name}

\item{x_vars}{vector of predictor column names}
}
\value{
final model created through backwards variable selection
}
\description{
Performs Backwards Variable Selection and returns created linear model
}
\examples{
backward_selection(mtcars, "mpg", colnames(mtcars)[2:11])

}
